separate CLI features for `pretty.js` (bin.js) with default features (pretty.js), to be able to use webpack with Node.js code in AWS Lambda environments
