+++ pretty.js
@@ -1,5 +1,3 @@
-#! /usr/bin/env node
-
 'use strict'
 
 var split = require('split2')
@@ -139,25 +137,3 @@ function pretty (opts) {
 }
 
 module.exports = pretty
-
-if (require.main === module) {
-  if (arg('-h') || arg('--help')) {
-    usage().pipe(process.stdout)
-  } else if (arg('-v') || arg('--version')) {
-    console.log(require('./package.json').version)
-  } else {
-    process.stdin.pipe(pretty({
-      timeTransOnly: arg('-t'),
-      levelFirst: arg('-l')
-    })).pipe(process.stdout)
-  }
-}
-
-function usage () {
-  return require('fs')
-    .createReadStream(require('path').join(__dirname, 'usage.txt'))
-}
-
-function arg (s) {
-  return !!~process.argv.indexOf(s)
-}
